put into hearing details object SSCS-11339
